board.ts refactor closure
